% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DIC_functions.R
\name{add_mle_param}
\alias{add_mle_param}
\title{Add MLE Parameter Values}
\usage{
add_mle_param(stan_fit, run_data)
}
\arguments{
\item{stan_fit}{stan fit object}

\item{run_data}{experimental data formated for stan}
}
\value{
run_data obejct with mle parameters included
}
\description{
Not sure yet.
}
