% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DIC_functions.R
\name{calc_DIC}
\alias{calc_DIC}
\title{Calculates DIC}
\usage{
calc_DIC(stan_fit)
}
\arguments{
\item{stan_fit}{stan fit object with \code{log_lik_t} field}
}
\value{
DIC value
}
\description{
Uses the samples of the log likelihoods in a stan fit file to calculate the
deviance information criterion (DIC).
}
