% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DIC_functions.R
\name{calc_DIC_hierarchical}
\alias{calc_DIC_hierarchical}
\title{Calculates DIC for Hierarchical Models}
\usage{
calc_DIC_hierarchical(stan_fit, stan_mle, run_data)
}
\arguments{
\item{stan_fit}{with \code{log_lik_t} field}
}
\value{
DIC value
}
\description{
Uses the
}
