% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_functions.R
\name{center_scale}
\alias{center_scale}
\title{Centers and scales pain rating in experimental data.}
\usage{
center_scale(exp_data)
}
\arguments{
\item{exp_data}{experimental data in the form of a dataframe object with fields (S,Q,Std,X,R).
S: subject id, Q: mean of presented que,
Std: standard deviation of presented que, X: administered stimulus, R: subject rating}
}
\value{
dataframe object from \code{exp_data} with subject ratings that are centered and scaled
}
\description{
Centers and scales pain rating in experimental data.
}
