% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_functions.R
\name{create_stan_data}
\alias{create_stan_data}
\title{Formats Data to Run with Stan Models}
\usage{
create_stan_data(exp_data)
}
\arguments{
\item{exp_data}{experimental data in the form of a dataframe object with fields (S,Q,Std,X,R).
S: subject id, Q: mean of presented que,
Std: standard deviation of presented que, X: administered stimulus, R: subject rating}
}
\value{
data to fit stan model
}
\description{
Formats Data to Run with Stan Models
}
