% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_functions.R
\name{fit_models}
\alias{fit_models}
\title{Fit Models}
\usage{
fit_models(models, stan_data, iter = 300, chains = 3)
}
\arguments{
\item{model_numbers}{array of numbers corresponding to the desired models}
}
\value{
an array of stan model objects corresponding to select models
}
\description{
Returns the stan model objects corresponding to the selected models from
the paper.
}
