% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_functions.R
\name{get_models}
\alias{get_models}
\title{Select Models from Hoskin2019}
\usage{
get_models(model_numbers)
}
\arguments{
\item{model_numbers}{array of numbers corresponding to the desired models}
}
\value{
an array of stan model objects corresponding to select models
}
\description{
Returns the stan model objects corresponding to the selected models from
the paper.
}
