% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DIC_functions.R
\name{get_param_mle}
\alias{get_param_mle}
\title{Find the Maximum Likelihood Estimate of Parameter}
\usage{
get_param_mle(stan_fit, parameter)
}
\arguments{
\item{stan_fit}{stan fit object}

\item{parameter}{parameter of interest}
}
\value{
the MLE of the parameter
}
\description{
Finds the Maximum Likelihood (MLE) of a given parameter in a stan_fit object
}
